refactor(middle): use int32 instead of int for userInfo
I've also update azutils repo for this commit

+++ internal/service/product_service.go
@@ -88,7 +88,7 @@ func (s *ProductService) ListSellerProducts(ctx context.Context, req *pb.ListSel
 		req.PageToken,
 		req.PageSize,
 		convertToBizProductStatus(req.ProductStatus),
-		int32(userID),
+		userID,
 		biz.UserRole(role),
 	)
 	if err != nil {
